Fix Level Zero loader build on Windows
On DLL platforms like Windows a shared library has the corresponding
import library (has .lib extension but not same as static library). This
import library is treated as ARCHIVE target. That is why ARCHIVE target
must be specified in cmake file so that cmake knows where to put import
file.

Signed-off-by: Gainullin <artur.gainullin@intel.com>