Wrap linker flags on Windows for IntelLLVM
The Intel C++ compiler requires linker flags to be wrapped, because
CMake passes them through the compiler driver.
Prefix Linker options with `${CMAKE_CXX_LINKER_WRAPPER_FLAG}`.
This variable is set to the option to pass a linker flag through the
compiler driver by CMake. For MSVC and clang-cl or IntelLLVM pre cmake
3.25 this will be empty, so the patch will work in these cases too.

Fixes: #205
Signed-off-by: Gergely Meszaros <gergely.meszaros@intel.com>