Add script to make a copy of headers in include/level_zero
This is to allow users consuming the headers from a
clone of the repository to be able to refer to the api headers
the same way as it would be if the level-zero headers were installed.
As in after running this script the following will compile
```c++
 #include <level_zero/ze_api.h>
```
when `<path-to-repo>/include` is part of the include directory.

Symlinks or alternatively wrapper header files could be added that would
forward to the previous location, but I stuck with copies, since the
files are generated anyway and any solution outside of full copies might
have corner cases. (Symlinks are not always supported on windows,
the build system might be grepping in the contents of the header etc.).

This PR also does not deprecate the current location of the headers.

The copies of the files will be checked in the next commit.

Signed-off-by: gergely.meszaros <gergely.meszaros@intel.com>